+++ lib/paw.dart
@@ -1,6 +1,9 @@
 library paw;
 
 export './src/interface.paw.dart';
+// core (logger and logInterface)
 export './src/logger.paw.dart';
-export './src/themes/dark.theme.paw.dart';
+// themes
+export './src/themes/themes.paw.dart';
+// log levels
 export './src/utils/log_levels.utils.paw.dart';
