+++ lib/src/ansi/ansi.paw.dart
@@ -1,0 +1,19 @@
+///
+/// ANSI Color Codes to beautify Logging
+///
+/// Paw try to provide a simple and effective way to apply ANSI color codes
+/// and styles to text in terminal or console applications. These ANSI codes are
+/// special character sequences that control the formatting, such as color and
+/// style, of text displayed in environments that support ANSI escape codes.
+///
+/// For more comprehensive details on ANSI colors, you can visit:
+/// [ANSI Colors Documentation](https://en.wikipedia.org/wiki/ANSI_escape_code)
+///
+/// To explore more ANSI color options, refer to:
+/// [Various ANSI Color Codes](https://talyian.github.io/ansicolors/)
+///
+library ansi;
+
+export './colors.ansi.paw.dart';
+export './styles.ansi.paw.dart';
+export './constants.ansi.paw.dart';
