+++ lib/src/themes/interface.themes.paw.dart
@@ -1,4 +1,4 @@
-import '../colors/ansi.paw.dart';
+import '../ansi/colors.ansi.paw.dart';
 
 ///
 /// `PawTheme` - An abstract class to help create custom themes to
